♻️  zb: Split handshake's cookies code into its own mod
